extract CustomPopover from Document
