+++ app/routes/document.tsx
@@ -20,19 +20,9 @@ import {
   TooltipTrigger,
 } from "~/components/ui/tooltip";
 import { Tweet } from "./tweet";
+import { CustomPopover } from "~/components/document/CustomPopover";
+import type { Annotation } from "~/types/types";
 
-type PopoverProps = {
-  docId: string;
-  docTitle: string;
-  selectionText: string;
-  annotationText: string;
-  setAnnotationText: (v: string) => void;
-  selectionRef: React.MutableRefObject<string>;
-  // optional: position; if you want to move with selection
-  x?: number;
-  y?: number;
-  onRequestClose: () => void;
-};
 
 // memo prevents unnecessary re-renders; most important is that the component
 // TYPE is stable by being top-level. Memo is a nice-to-have.
@@ -111,15 +101,7 @@ function NotePopover({
 
 type LoaderData = {
   document: { id: string; content: string, title: string };
-  annotations: Array<{
-    id: string;
-    start: number;
-    end: number;
-    quote: string;
-    note: string;
-    prefix: string;
-    suffix: string;
-  }>;
+  annotations: Annotation[];
 };
 
 export async function loader({
@@ -142,168 +124,6 @@ export async function loader({
 }
 
 export default function Document() {
-  const { id } = useParams();
-
-
-
-  const CustomPopover = memo(function CustomPopover({
-    docId,
-    docTitle,
-    selectionText,
-    annotationText,
-    setAnnotationText,
-    selectionRef,
-    onRequestClose,
-    x = 0,
-    y = 0,
-  }: PopoverProps) {
-    const rootRef = useRef<HTMLDivElement>(null);
-    const hiddenRef = useRef<HTMLInputElement>(null);
-    const noteRef = useRef<HTMLTextAreaElement>(null);
-
-    useEffect(() => {
-      const handlePointerDown = (e: PointerEvent) => {
-        const el = rootRef.current;
-        if (!el) return;
-        // If click is outside the popover, close it
-        if (!el.contains(e.target as Node)) {
-          onRequestClose();
-        }
-      };
-      // capture = true so we see the event even if inner handlers stopPropagation
-      window.document.addEventListener("pointerdown", handlePointerDown, true);
-      return () =>
-        window.document.removeEventListener("pointerdown", handlePointerDown, true);
-    }, [onRequestClose]);
-    const onSubmit: React.FormEventHandler<HTMLFormElement> = (e) => {
-      let parsed: any = null;
-      try {
-        parsed = JSON.parse(selectionRef.current);
-      } catch { }
-      if (!parsed) {
-        e.preventDefault();
-        return;
-      }
-
-      const payload = {
-        documentId: docId,
-        start: parsed.start,
-        end: parsed.end,
-        quote: parsed.quote,
-        prefix: parsed.prefix,
-        suffix: parsed.suffix,
-        body: noteRef.current?.value ?? "",
-      };
-
-      if (hiddenRef.current) hiddenRef.current.value = JSON.stringify(payload);
-    };
-
-    const [tweetSidenote, setTweetSidenote] = useState("");
-
-    return (
-      <div
-        ref={rootRef}
-        style={{
-          position: "fixed",
-          left: x,
-          top: y,
-          background: "white",
-          border: "1px solid #e5e7eb",
-          padding: "16px 20px",
-          borderRadius: 12,
-          boxShadow: "0 8px 32px rgba(0,0,0,0.18)",
-          zIndex: 1,
-          minWidth: 320,
-          maxWidth: 500,
-          pointerEvents: "auto",
-        }}
-      >
-        <p className="mb-3 text-sm text-gray-700 font-medium break-words">
-          {selectionText}
-        </p>
-        <div className="flex flex-col gap-3 items-center">
-          <Form
-            className="flex w-full items-end"
-            method="post"
-            action={`/workspace/document/${docId}/save-annotation`}
-            onSubmit={onSubmit}
-          >
-            <input ref={hiddenRef} type="hidden" name="annotation" />
-            <textarea
-              ref={noteRef}
-              name="note"
-              placeholder="Type text..."
-              onChange={(e) => {
-                setTweetSidenote(e.target.value);
-                e.currentTarget.style.height = "auto";
-                e.currentTarget.style.height = `${Math.min(e.currentTarget.scrollHeight, 3 * 24)}px`; // 3 lines max (assuming 24px line-height)
-              }}
-              rows={1}
-              className="w-full resize-none overflow-y-auto bg-transparent border p-1 rounded-sm focus:ring-0 focus:outline-none leading-6"
-              onMouseDown={(e) => e.stopPropagation()}
-            />
-            <TooltipProvider>
-              <Tooltip>
-                <TooltipTrigger asChild>
-                  <Button
-                    size="icon"
-                    variant="ghost"
-                    className="ml-1"
-                    type="submit"
-
-                  >
-                    <CornerDownLeft className="h-4 w-4" />
-                  </Button>
-                </TooltipTrigger>
-                <TooltipContent>Add annotation</TooltipContent>
-              </Tooltip>
-            </TooltipProvider>
-          </Form>
-          <div className="flex flex-row">
-            <Form method="post" action={`/workspace/document/${docId}/chat-create`}>
-              <Tooltip>
-                <TooltipTrigger asChild>
-                  <Button size="icon" variant="ghost" onClick={() => { }}>
-                    <MessageCirclePlus className="h-2 w-2" />
-                    <span className="sr-only">Create new chat</span>
-                  </Button>
-                </TooltipTrigger>
-                <TooltipContent>
-                  <p>Create new chat</p>
-                </TooltipContent>
-              </Tooltip>
-            </Form>
-
-            <Tooltip>
-              <TooltipTrigger asChild>
-                <Button size="icon" variant="ghost" onClick={() => {
-                  setIncludeSelection(true);
-                }}>
-                  <MessageSquareReply className="h-2 w-2" />
-                  <span className="sr-only">Add to existing chat</span>
-                </Button>
-              </TooltipTrigger>
-              <TooltipContent>
-                <p>Add to existing chat</p>
-              </TooltipContent>
-            </Tooltip>
-            <Tooltip>
-              <TooltipTrigger asChild>
-                <Button size="icon" variant="ghost" onClick={() => {
-                }}>
-                  <Tweet title={docTitle} annotationText={tweetSidenote} selectionText={selectionText} docId={docId}></Tweet>
-                </Button>
-              </TooltipTrigger>
-              <TooltipContent>
-                <p>Tweet annotation</p>
-              </TooltipContent>
-            </Tooltip>
-          </div>
-        </div>
-      </div>
-    );
-  });
-
 
   const [notePopup, setNotePopup] = useState<null | {
     x: number;
@@ -320,6 +140,7 @@ export default function Document() {
     }>();
 
   const { document, annotations } = useLoaderData() as LoaderData;
+  const id = document.id
   const docContent = () => {
     return { __html: document.content };
   };
@@ -471,6 +292,7 @@ export default function Document() {
             annotationText={annotationText}
             setAnnotationText={setannotationText}
             selectionRef={selectionRef}
+            setIncludeSelection={setIncludeSelection}
             x={popup.x}
             y={popup.y}
             onRequestClose={() => setPopup(null)}
